Faster distance computation: skip missing features
For each training point, record the indices of the features that were defined
for that point. Pass these values as a slice to the distance function, so that
it only tries to compute distances between the training point and all other
points in the testing set for features defined in the training point.

While we still have to test that features are defined for each of the other
points in the testing set, we save significant computation by not having to test
for every single distance computation whether the feature is defined for the
point being trained on.

On my Intel(R) Core(TM) i7-5600U CPU @ 2.60GHz processor, the attack now runs in
75% of time that it used to.

% time ./precompute
...
2016/09/25 14:10:12 starting to learn distance...
        distance... 5999 (0-6000)
2016/09/25 14:13:03 finished
2016/09/25 14:13:03 started computing accuracy...
        accuracy... 11999 (0-12000)
2016/09/25 14:27:19 finished
2016/09/25 14:27:19 Accuracy: 0.806667 0.993667
./precompute  993.92s user 2.07s system 96% cpu 17:09.93 total

% time ./knn.fixed
...
2016/09/25 14:10:12 starting to learn distance...
        distance... 5999 (0-6000)
2016/09/25 14:14:26 finished
2016/09/25 14:14:26 started computing accuracy...
        accuracy... 11999 (0-12000)k
2016/09/25 14:32:24 finished
2016/09/25 14:32:24 Accuracy: 0.806667 0.993667
./knn.fixed  1296.77s user 2.22s system 97% cpu 22:14.45 total